[GTK4] Do not unparent the model of a disposed drop-down menu
On GTK4 the handle of a DROP_DOWN menu is its GMenu model, so disposing
the menu passed it to gtk_widget_unparent, which fails with a critical.
Also release the model, which was leaked for every menu.

+++ bundles/org.eclipse.swt/Eclipse SWT/gtk/org/eclipse/swt/widgets/Menu.java
@@ -1726,10 +1726,19 @@ void releaseWidget () {
  */
 @Override
 void destroyWidget () {
-	super.destroyWidget();
+	if (GTK.GTK4 && (style & SWT.DROP_DOWN) != 0) {
+		/* The handle of a GTK4 drop-down menu is its GMenu model, not a widget */
+		releaseHandle ();
+	} else {
+		super.destroyWidget();
+	}
 	if (menuHandle != 0) {
 		OS.g_object_unref(menuHandle);
 		menuHandle = 0;
+	}
+	if (modelHandle != 0) {
+		OS.g_object_unref(modelHandle);
+		modelHandle = 0;
 	}
 }
